+++ Sources/PointFree/Admin/Index.swift
@@ -7,6 +7,7 @@ import PointFreeRouter
 import PointFreePrelude
 import Prelude
 import Tuple
+import Views
 
 public let adminEmails: [EmailAddress] = [
   "mbw234@gmail.com",
